requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
distro==1.9.0
Flask==3.1.2
frozenlist==1.6.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
magic-filter==1.0.12
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.4.3
openai==1.78.1
propcache==0.3.1
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
Pyrogram==2.0.106
PySocks==1.7.1
pyTelegramBotAPI==4.29.1
python-dotenv==1.1.0
python-telegram-bot==22.0
PyYAML==6.0.2
requests==2.32.5
rich==14.0.0
rsa==4.9.1
sniffio==1.3.1
Telethon==1.40.0
TgCrypto==1.2.5
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.0
